+++ meridian/analysis/summarizer.py
@@ -77,6 +77,7 @@ class Summarizer:
       filepath: str,
       start_date: tc.Date = None,
       end_date: tc.Date = None,
+      use_kpi: bool = False,
   ):
     """Generates and saves the HTML results summary output.
 
@@ -86,15 +87,18 @@ class Summarizer:
       start_date: Optional start date selector, *inclusive*, in _yyyy-mm-dd_
         format.
       end_date: Optional end date selector, *inclusive* in _yyyy-mm-dd_ format.
+      use_kpi: If `True`, calculate the incremental KPI. Otherwise, calculate
+        the incremental revenue using the revenue per KPI (if available).
     """
     os.makedirs(filepath, exist_ok=True)
     with open(os.path.join(filepath, filename), 'w') as f:
-      f.write(self._gen_model_results_summary(start_date, end_date))
+      f.write(self._gen_model_results_summary(start_date, end_date, use_kpi))
 
   def _gen_model_results_summary(
       self,
       start_date: tc.Date = None,
       end_date: tc.Date = None,
+      use_kpi: bool = False,
   ) -> str:
     """Generate HTML results summary output (as sanitized content str)."""
     all_dates = self._meridian.input_data.time_coordinates.all_dates
@@ -140,6 +144,7 @@ class Summarizer:
     cards_htmls = self._create_cards_htmls(
         template_env,
         selected_times=selected_times,
+        use_kpi=use_kpi,
     )
 
     return html_template.render(
@@ -150,6 +155,7 @@ class Summarizer:
       self,
       template_env: jinja2.Environment,
       selected_times: Sequence[str] | None,
+      use_kpi: bool,
   ) -> Sequence[str]:
     """Creates the HTML snippets for cards in the summary page."""
     media_summary = visualizer.MediaSummary(
@@ -165,10 +171,13 @@ class Summarizer:
     )
     cards = [
         self._create_model_fit_card_html(
-            template_env, selected_times=selected_times
+            template_env, selected_times=selected_times, use_kpi=use_kpi
         ),
         self._create_outcome_contrib_card_html(
-            template_env, media_summary, selected_times=selected_times
+            template_env,
+            media_summary,
+            selected_times=selected_times,
+            use_kpi=use_kpi,
         ),
         self._create_performance_breakdown_card_html(
             template_env, media_summary
@@ -179,16 +188,17 @@ class Summarizer:
             media_summary=media_summary,
             media_effects=media_effects,
             reach_frequency=reach_frequency,
+            use_kpi=use_kpi,
         ),
     ]
     return cards
 
   def _create_model_fit_card_html(
-      self, template_env: jinja2.Environment, **kwargs
+      self, template_env: jinja2.Environment, use_kpi: bool, **kwargs
   ) -> str:
     """Creates the HTML snippet for the Model Fit card."""
     model_fit = self._model_fit
-    outcome = self._kpi_or_revenue()
+    outcome = self._kpi_or_revenue(use_kpi)
     expected_actual_outcome_chart = formatter.ChartSpec(
         id=summary_text.EXPECTED_ACTUAL_OUTCOME_CHART_ID,
         description=summary_text.EXPECTED_ACTUAL_OUTCOME_CHART_DESCRIPTION_FORMAT.format(
@@ -197,7 +207,9 @@ class Summarizer:
         chart_json=model_fit.plot_model_fit(**kwargs).to_json(),
     )
 
-    predictive_accuracy_table = self._predictive_accuracy_table_spec(**kwargs)
+    predictive_accuracy_table = self._predictive_accuracy_table_spec(
+        use_kpi=use_kpi, **kwargs
+    )
     insights = summary_text.MODEL_FIT_INSIGHTS_FORMAT
 
     return formatter.create_card_html(
@@ -207,9 +219,11 @@ class Summarizer:
         [expected_actual_outcome_chart, predictive_accuracy_table],
     )
 
-  def _predictive_accuracy_table_spec(self, **kwargs) -> formatter.TableSpec:
+  def _predictive_accuracy_table_spec(
+      self, use_kpi: bool, **kwargs
+  ) -> formatter.TableSpec:
     """Creates the HTML snippet for the predictive accuracy table."""
-    outcome = self._kpi_or_revenue()
+    outcome = self._kpi_or_revenue(use_kpi)
     model_diag = self._model_diagnostics
     table = model_diag.predictive_accuracy_table(column_var=c.METRIC, **kwargs)
 
@@ -270,9 +284,10 @@ class Summarizer:
       template_env: jinja2.Environment,
       media_summary: visualizer.MediaSummary,
       selected_times: Sequence[str] | None,
+      use_kpi: bool,
   ) -> str:
     """Creates the HTML snippet for the Outcome Contrib card."""
-    outcome = self._kpi_or_revenue()
+    outcome = self._kpi_or_revenue(use_kpi)
 
     num_selected_times = (
         self._meridian.n_times
@@ -442,9 +457,10 @@ class Summarizer:
       media_summary: visualizer.MediaSummary,
       media_effects: visualizer.MediaEffects,
       reach_frequency: visualizer.ReachAndFrequency | None,
+      use_kpi: bool,
   ) -> str:
     """Creates the HTML snippet for the Optimal Analyst card."""
-    outcome = self._kpi_or_revenue()
+    outcome = self._kpi_or_revenue(use_kpi)
     charts = []
     charts.append(
         formatter.ChartSpec(
@@ -457,6 +473,7 @@ class Summarizer:
                 selected_times=(
                     frozenset(selected_times) if selected_times else None
                 ),
+                use_kpi=use_kpi,
                 plot_separately=False,
                 include_ci=False,
                 num_channels_displayed=7,
@@ -524,9 +541,7 @@ class Summarizer:
         rf_channel=most_spend_rf_channel
     ).optimal_frequency
 
-  def _kpi_or_revenue(self) -> str:
-    if self._meridian.input_data.revenue_per_kpi is not None:
-      outcome_str = c.REVENUE
-    else:
-      outcome_str = c.KPI.upper()
-    return outcome_str
+  def _kpi_or_revenue(self, use_kpi: bool) -> str:
+    if use_kpi or self._meridian.input_data.revenue_per_kpi is None:
+      return c.KPI.upper()
+    return c.REVENUE